+++ packages/state/src/index.ts
@@ -1,17 +1,5 @@
-// Export walletProvider types
-export {
-  ChainName,
-  ChainMetadata,
-  Connector,
-  InstructionStepName,
-  NetworkDetails,
-  WalletChain,
-} from "./providers/walletProvider/types";
-
-export { useSorobanReact } from "./providers/walletProvider/hooks/useSorobanReact";
-
 // Export freighter
-export { freighter } from "./providers/walletProvider/freighter";
+export { freighter } from "./state/wallet/freighter";
 
 // Export chains
 export {
@@ -20,17 +8,10 @@ export {
   testnet,
   sandbox,
   standalone,
-} from "./providers/walletProvider/chains";
+  allChains,
+} from "./state/wallet/chains";
 
-export { useLoadToken } from "./hooks/useLoadToken";
-export { usePairInfos } from "./hooks/usePairInfos";
+export * from "./state/store";
 
-export { useContractSign } from "./hooks/contracts/useContractSign";
-
-// Export walletProvider
-export { WalletProvider } from "./providers/walletProvider";
-
-// Export PhoenixProvider
-export { PhoenixProvider } from "./providers/phoenixProvider";
-
-export { contractTransaction } from "./hooks/contracts/useContractSign";
+export * from "./soroban";
+export * from "./clients";