+++ CHANGELOG.md
@@ -8,5 +8,23 @@ and this project adheres to [Semantic Versioning].
 ## [Unreleased]
 
 ### Added
+- Initial commits. [general]
+- Added packages/ui, packages/state, packages/demo, packages/utils, packages/core, packages/demo. [general]
+- Added some example React Next code to the demo package for testing purposes. [demo]
+- Added button component. [ui]
+- Added storybook. [ui]
 
-- Initial UI package creation ([#1](https://github.com/Phoenix-Protocol-Group/phoenix-frontend/pull/1]))
+### Removed
+- Removed Soroban React and replaced with Zustand for state management. [state]
+
+### Changed
+- Migrated state management to use Zustand instead of Soroban React. [state]
+
+### Deprecated
+- None.
+
+### Fixed
+- None.
+
+### Security
+- None.